+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Master
+## 3.15.4 (2019-07-23)
 
 - Check if it is a valid JSON object in `-JSONString`.
 - Added `-JSONData`, `-JSONDataWithOptions:` to `NSArray`, `NSDictionary`.